CATROID-924 The app crashes on pressing the undo button when you switch back to the looks tab (from the scripts or sounds tab) after editing a look! The app should hide the undo button!

+++ catroid/src/main/java/org/catrobat/catroid/ui/SpriteActivity.java
@@ -237,7 +237,7 @@ public class SpriteActivity extends BaseActivity {
 			setUndoMenuItemVisibility(false);
 			showUndoMenuItem(isUndoMenuItemVisible);
 			Fragment fragment = getCurrentFragment();
-			if (fragment instanceof LookListFragment && !((LookListFragment) fragment).undo()) {
+			if (fragment instanceof LookListFragment && !((LookListFragment) fragment).undo() && currentLookData != null) {
 				((LookListFragment) fragment).deleteItem(currentLookData);
 				currentLookData.dispose();
 				currentLookData = null;
